requirements.txt
requests
bs4
selenium
openai
python-dotenv